`Refactor StarRating component to use new Star component and fix import path`
