+++ src/starRating/StarRating.js
